add toolbar toggles, and split model classes into separate files
