+++ projects/media-viewer/src/lib/media-viewer/toolbar/left-pane/left-pane.component.html
@@ -1,25 +1,28 @@
 <div id="toolbarViewerLeft">
-  <button id="sidebarToggle" class="toolbarButton" title="Toggle Sidebar" tabindex="11"
+  <button *ngIf="toolbarToggles.showSidebarToggleBtn | async"
+          id="sidebarToggle" class="toolbarButton" title="Toggle Sidebar" tabindex="11"
           data-l10n-id="toggle_sidebar" (click)="toggleSideBar()">
     <span data-l10n-id="toggle_sidebar_label">Toggle Sidebar</span>
   </button>
   <div class="toolbarButtonSpacer"></div>
-  <button id="viewFind" class="toolbarButton" title="Find in Document" tabindex="12" data-l10n-id="findbar"
+  <button *ngIf="toolbarToggles.showSearchbarToggleBtn | async" id="viewFind" class="toolbarButton" title="Find in Document" tabindex="12" data-l10n-id="findbar"
           (click)="toggleSearchBar()">
     <span data-l10n-id="findbar_label">Find</span>
   </button>
-  <div class="splitToolbarButton hiddenSmallView">
-    <button class="toolbarButton pageUp" title="Previous Page" id="previous" tabindex="13" (click)="increasePageNumber()"
-            data-l10n-id="previous">
-      <span data-l10n-id="previous_label">Previous</span>
-    </button>
-    <div class="splitToolbarButtonSeparator"></div>
-    <button class="toolbarButton pageDown" title="Next Page" id="next" tabindex="14" data-l10n-id="next" (click)="decreasePageNumber()">
-      <span data-l10n-id="next_label">Next</span>
-    </button>
-  </div>
-  <input type="number" id="pageNumber" class="toolbarField pageNumber" title="Page" value="1" size="4" min="1"
-         [value]="pageNumber"
-         tabindex="15" data-l10n-id="page" (change)="setCurrentPageNumber(pageNumberInput.value)" #pageNumberInput>
-  <span id="numPages" class="toolbarLabel"></span>
+  <ng-container *ngIf="toolbarToggles.showNavigationBtns | async">
+    <div class="splitToolbarButton hiddenSmallView">
+      <button class="toolbarButton pageUp" title="Previous Page" id="previous" tabindex="13" (click)="increasePageNumber()"
+              data-l10n-id="previous">
+        <span data-l10n-id="previous_label">Previous</span>
+      </button>
+      <div class="splitToolbarButtonSeparator"></div>
+      <button class="toolbarButton pageDown" title="Next Page" id="next" tabindex="14" data-l10n-id="next" (click)="decreasePageNumber()">
+        <span data-l10n-id="next_label">Next</span>
+      </button>
+    </div>
+    <input type="number" id="pageNumber" class="toolbarField pageNumber" title="Page" value="1" size="4" min="1"
+           [value]="pageNumber"
+           tabindex="15" data-l10n-id="page" (change)="setCurrentPageNumber(pageNumberInput.value)" #pageNumberInput>
+    <span id="numPages" class="toolbarLabel"></span>
+  </ng-container>
 </div>
