+++ projects/media-viewer/src/lib/media-viewer/toolbar/middle-pane/middle-pane.component.html
@@ -1,4 +1,4 @@
-<div id="toolbarViewerMiddle">
+<div id="toolbarViewerMiddle" *ngIf="showZoomBtns">
   <div class="splitToolbarButton">
     <button id="zoomOut" class="toolbarButton zoomOut" title="Zoom Out" tabindex="21" data-l10n-id="zoom_out"
             (click)="stepZoom(-0.2)">
@@ -11,28 +11,28 @@
     </button>
   </div>
   <span id="scaleSelectContainer" class="dropdownToolbarButton">
-        <select id="scaleSelect" title="Zoom" tabindex="23" data-l10n-id="zoom" (change)="zoom($event.target.value)">
-          <option #zoomSelect id="customScaleOption" title="" [value]="zoomValue.value">{{zoomValue.value * 100 | number: '1.0-0'}}%</option>
-          <option title="" value="0.1" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 10 }'>10%</option>
-          <option title="" value="0.25" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 25 }'>25%</option>
-          <option title="" value="0.5" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 50 }'>50%</option>
-          <option title="" value="0.75" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 75 }'>75%</option>
-          <option title="" value="1" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 100 }'>100%</option>
-          <option title="" value="1.25" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 125 }'>125%</option>
-          <option title="" value="1.5" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 150 }'>150%</option>
-          <option title="" value="2.5" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 250 }'>250%</option>
-          <option title="" value="3" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 300 }'>300%</option>
-          <option title="" value="5" data-l10n-id="page_scale_percent"
-                  data-l10n-args='{ "scale": 500 }'>500%</option>
-        </select>
-      </span>
+      <select id="scaleSelect" title="Zoom" tabindex="23" data-l10n-id="zoom" (change)="zoom($event.target.value)">
+        <option #zoomSelect id="customScaleOption" title="" [value]="zoomValue.value">{{zoomValue.value * 100 | number: '1.0-0'}}%</option>
+        <option title="" value="0.1" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 10 }'>10%</option>
+        <option title="" value="0.25" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 25 }'>25%</option>
+        <option title="" value="0.5" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 50 }'>50%</option>
+        <option title="" value="0.75" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 75 }'>75%</option>
+        <option title="" value="1" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 100 }'>100%</option>
+        <option title="" value="1.25" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 125 }'>125%</option>
+        <option title="" value="1.5" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 150 }'>150%</option>
+        <option title="" value="2.5" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 250 }'>250%</option>
+        <option title="" value="3" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 300 }'>300%</option>
+        <option title="" value="5" data-l10n-id="page_scale_percent"
+                data-l10n-args='{ "scale": 500 }'>500%</option>
+      </select>
+  </span>
 </div>
