+++ projects/media-viewer/src/lib/media-viewer/toolbar/middle-pane/middle-pane.component.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, Input, ViewChild } from '@angular/core';
 import { Subject } from 'rxjs';
-import { ZoomOperation, StepZoomOperation, ZoomValue } from '../../media-viewer.model';
+import { StepZoomOperation, ZoomOperation, ZoomValue } from '../../model/viewer-operations';
 
 @Component({
   selector: 'mv-tb-middle-pane',
@@ -12,6 +12,8 @@ export class ToolbarMiddlePaneComponent {
   @Input() zoomEvent: Subject<ZoomOperation>;
   @Input() stepZoomEvent: Subject<StepZoomOperation>;
   @Input() zoomValue: ZoomValue;
+  @Input() showZoomBtns: boolean;
+
   @ViewChild('zoomSelect') zoomSelect: ElementRef;
 
   constructor() {}
