+++ projects/media-viewer/src/lib/media-viewer/toolbar/sub-toolbar/sub-toolbar.component.spec.ts
@@ -1,11 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { SubToolbarComponent } from './sub-toolbar.component';
-import {
-  ActionEvents,
-  DownloadOperation,
-  PrintOperation,
-  RotateOperation
-} from '../../media-viewer.model';
+import { ActionEvents } from '../../model/action-events';
+import { DownloadOperation, PrintOperation, RotateOperation } from '../../model/viewer-operations';
 
 describe('SubToolbarComponent', () => {
   let component: SubToolbarComponent;
