+++ projects/media-viewer/src/public_api.ts
@@ -4,4 +4,4 @@
 
 export * from './lib/media-viewer.module';
 export * from './lib/media-viewer/toolbar/toolbar.module';
-export * from './lib/media-viewer/media-viewer.model';
+export * from './lib/media-viewer/model/action-events';
